requirements.txt
# StudyMate - AI-Powered Academic PDF RAG System
# Core dependencies with compatible versions

# Web Framework
streamlit>=1.28.0

# PDF Processing (using more compatible version)
PyMuPDF>=1.23.0

# Machine Learning & Embeddings
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0

# IBM Watsonx Integration
ibm-watsonx-ai>=1.0.0

# Hugging Face Integration
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0

# OpenAI Integration
openai>=1.0.0

# Google Gemini Integration
google-generativeai>=0.3.0

# Environment Management
python-dotenv>=1.0.0


# Additional utilities
pandas>=2.0.0
requests>=2.31.0

# OCR dependencies for image-to-text
pytesseract
opencv-python

# Document Generation
python-docx>=0.8.11
reportlab>=4.0.0

# Voice Assistant
SpeechRecognition>=3.10.0
pyaudio>=0.2.11
pydub>=0.25.1
